Community to EAP conversion -- Do not ignore tests that parses legacy configurations
In EAP, all the legacy configurations we are currently testing are for EAP base, so parsers for EAP or XP should be able to test them.
In this commit, we disable the check that verifies if we are using a EAP or XP server to verify these legacy configuration files.

Signed-off-by: Yeray Borges <yborgess@redhat.com>